rename : 일부 태그들 컴포넌트로 분리 및 이름변경
