+++ src/main/front/src/components/Contents/ItemUser.js
@@ -1,11 +1,11 @@
 import React from "react";
 
-import styles from "../../styles/Contents.module.css";
+import styles from "../../styles/ItemUser.module.css";
 
 function ItemUser(props) {
   return (
     <div className={styles.itemUser}>
-      <div className={styles.userProfile} />
+      <div className={styles.pic} />
       <h4 className={styles.userName}>{props.userId}</h4>
     </div>
   );
