+++ src/main/front/src/components/Modal/ProfileModal.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import styles from "../styles/Header.module.css";
+import styles from "../../styles/Header.module.css";
 
 function ProfileModal(props) {
   return (
