--- src/main/front/src/styles/ContentDetail.module.css
@@ -1,45 +1,0 @@
-.bg {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.6);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.closeBtn {
-  position: fixed;
-  top: 20px;
-  right: 20px;
-  font-size: 20px;
-  color: #fff;
-}
-
-.closeBtn:hover {
-  cursor: pointer;
-}
-
-/* modal 창*/
-.body {
-  position: absolute;
-  max-width: 1200px;
-  max-height: 90vh;
-  width: 85vw;
-  height: 50vw;
-  text-align: center;
-  background-color: #fff;
-  border-radius: 5px;
-  box-shadow: 0 0 12px rgba(0, 0, 0, 0.1);
-
-  display: flex;
-}
-
-.mainPic {
-  height: 100%;
-  width: 60%;
-  min-width: 30%;
-  background-color: pink;
-}
