+++ src/main/front/src/styles/Contents.module.css
@@ -6,33 +6,15 @@
   padding: 30px 0;
 }
 
-.contentsItem {
+.Item {
   margin-bottom: 30px;
   border: 1px solid rgba(0, 0, 0, 0.2);
   border-radius: 7px;
 }
 
-/* itemUser*/
-.itemUser {
-  padding: 0 12px;
-  display: flex;
-  align-items: center;
-  padding: 8px 16px;
-}
-
-.itemUser .userProfile {
-  width: 50px;
-  height: 50px;
-  background-color: antiquewhite;
-  border-radius: 50%;
-}
-
-.itemUser .userName {
-  margin-left: 12px;
-}
-
 /*itemInfo*/
 .itemInfo {
+  border-top: 1px solid rgba(0, 0, 0, 0.2);
   padding: 8px 16px;
 }
