+++ src/main/front/src/styles/ItemUser.module.css
@@ -1,0 +1,19 @@
+/* itemUser*/
+.itemUser {
+  padding: 0 12px;
+  display: flex;
+  align-items: center;
+  padding: 8px 16px;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.08);
+}
+
+.pic {
+  width: 50px;
+  height: 50px;
+  background-color: antiquewhite;
+  border-radius: 50%;
+}
+
+.userName {
+  margin-left: 12px;
+}
